requirements.txt
# SQL Autograder Dependencies

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Google Gemini API
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.7.0